requirements.txt
PyYAML==3.11
docutils==0.12
matplotlib==1.4.2
numpy==1.8.2
pybedtools==0.7.5
pytabix==0.1
requests==2.4.3
snakemake==3.6.1
begins==0.9
